% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{A}
\alias{A}
\title{Threshold parameters for genes.}
\usage{
A(net)
}
\arguments{
\item{net}{The network to be transformed.}
}
\value{
The network with associated parameters.
}
\description{
Associates with each regulatory set of genes a vector
  determining whether said genes have a positive or negative effect.
}
\examples{
\dontrun{A(net)}
}
