% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{aToFunction}
\alias{aToFunction}
\title{Boolean functions from threshold parameters.}
\usage{
aToFunction(net)
}
\arguments{
\item{net}{A network which has threshold parameters.}
}
\value{
A Network with logic functions associated.
}
\examples{
\dontrun{aToFunction(net)}
}
