% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{bestA}
\alias{bestA}
\title{Jointly optimize all threshold parameters.}
\usage{
bestA(ts.multi, net, g, i)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{i}{The index of the regulatory function.}
}
\value{
A vector containing the best threshold parameters found.
}
\description{
Wrapping \code{\link{bestPartialA}}, this doesn't hold any values constant.
}
\seealso{
\code{\link{bestPartialA}}
}
