% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{bestC}
\alias{bestC}
\title{Find optimal probabilities for a set of regulatory functions.}
\usage{
bestC(ts.multi, net, g)
}
\arguments{
\item{ts.multi}{List of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}
}
\value{
A vector containing the probabilities optimizing the loss function.
}
\description{
Using standard optimization \code{solnp}, this tries to find the
probabilities for a single gene which minimizes the loss function.
}
\seealso{
\code{\link{optimizeC}}
}
