% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{bestPartialA}
\alias{bestPartialA}
\title{Optimize a subset of threshold parameters.}
\usage{
bestPartialA(ts.multi, net, g, i, subset)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{i}{The index of the specific regulatory function for the optimization.}

\item{subset}{The subset of parameters over which to optimize.}
}
\value{
A vector containing the best threshold parameters found.
}
\description{
While holding everything but the subset constant, optimize over this subset.
This essentially just tries every possible combination for the values on
this subset. However, zeroes are not allowed, as they have been found never
to come up while making the inference much, much, slower.
}
\seealso{
\code{\link{bestA}}
}
