% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{createNetwork}
\alias{createNetwork}
\title{Create network.}
\usage{
createNetwork(inputProbabilities = 1, ...)
}
\arguments{
\item{inputProbabilities}{List of probabilities determining how many
regulatory sets there are per gene.}

\item{...}{Other parameters, used by BoolNet.}
}
\value{
A network.
}
\examples{
\dontrun{
createNetwork(c(0.5, 0.5))
}
}
