% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{FullRun}
\alias{FullRun}
\title{Complete run of network generation and inference.}
\usage{
FullRun(n = 20, k = 5, p = 0.01, num.timepoints = 10,
  num.experiments = 50, topology = "homogeneous", gamma = 2.5,
  n.cores = detectCores() - 1, seed = 111, partial = FALSE,
  verbal = FALSE)
}
\arguments{
\item{n}{Size of the network.}

\item{k}{The number of inputs per regulatory function for each gene, if homogeneous
topology is used}

\item{p}{The probability of a perturbation.}

\item{num.timepoints}{The number of time points per timeseries generated.}

\item{num.experiments}{The number of timeseries to generate.}

\item{topology}{The topology to be used. Can be "homogeneous" or "scale_free".}

\item{gamma}{The exponent for the power law if topology = "scale_free".}

\item{n.cores}{The number of cores to use in the inference.}

\item{seed}{The random seed to use.}

\item{partial}{If TRUE, a network using partial optimization should be inferred. Defaults to FALSE.}

\item{verbal}{If TRUE, show progress as to which genes are currently being worked on.}
}
\description{
Automatically generate a network, generate timeseries data from it, and do inference.
Saves the inferred and true networks to a subdirectory.
}
\seealso{
\code{\link{setupPBN}} \code{\link{inferPBN}}
}
