% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-PBN.R
\name{inferPBN}
\alias{inferPBN}
\title{Infer a Threshold PBN from a list of timeseries.}
\usage{
inferPBN(ts.multi, p = 0.01, L.max = 2, n.cores = detectCores() - 1,
  seed = 111, partial = FALSE, verbal = FALSE)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{p}{The noise probability to use for the network.}

\item{L.max}{The maximum number of regulatory networks to try to infer.}

\item{n.cores}{The number of cores to be used.}

\item{seed}{The random seed to use.}

\item{partial}{Whether or not to *also* infer a network where only partial
optimization over the last threshold parameter is performed. Defaults to
  FALSE.}

\item{verbal}{Show progress report? Defaults to FALSE.}
}
\value{
A list containing the inferred network and the time taken to perform
  the inference.
}
\description{
This is a wrapper for the whole inference procedure and makes use of
\code{\link{innovateGeneUntilSaturated}}.
For every gene it infers regulatory functions until the upper bound is
reached or there is no more improvement from adding more functions.
}
\examples{
net.true <- createNetwork(inputProbabilities=c(0.5, 0.5), n=5, k=2)
ts.multi <- simulateNetwork(net.true, 10, 20)
inferred.list <- inferPBN(ts.multi, n.cores=1)
net.inferred <- inferred.list$net.inferred
time <- inferred.list$time.complete
}
\seealso{
\code{\link{FullRun}}
}
