% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{innovateGene}
\alias{innovateGene}
\title{Greedily add new gene to regulatory function.}
\usage{
innovateGene(ts.multi, net, g, partial = FALSE, verbal = TRUE)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{partial}{Whether only the most recently added gene should have its
parameter optimized. Defaults to FALSE.}

\item{verbal}{Whether messages to indicate progress should be shown.
Defaults to TRUE.}
}
\value{
A network.
}
\description{
des
}
\details{
Try to add regulators to the "most recent" regulatory set. Stop only when
there is no improvement any longer.
The choice of the gene to add is done greedily.
}
