% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{innovateGeneUntilSaturated}
\alias{innovateGeneUntilSaturated}
\title{Add new genes to regulatory set until no more improvement is made.}
\usage{
innovateGeneUntilSaturated(ts.multi, net, g, partial = FALSE, verbal = TRUE,
  thresh = 0.1, k.max = 3)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{partial}{Whether optimization of threshold parameter a is done only
for latest gene. Defaults to FALSE.}

\item{verbal}{Whether progress messages should be displayed.}

\item{thresh}{By how much the improvement has to be to count. This has to be
greater than zero as otherwise new genes will always be added to the brim.}

\item{k.max}{How many regulatory genes to allow maximally. To avoid
excessive computation, this should be set to a small constant. Defaults to
3.}
}
\value{
A network.
}
\description{
des
}
\details{
det
}
