% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{Loss}
\alias{Loss}
\title{Compute loss of the network on the timeseries.}
\usage{
Loss(ts.multi, net, prior = TRUE)
}
\arguments{
\item{ts.multi}{List of timeseries.}

\item{net}{A PBN.}

\item{prior}{If TRUE, loss for the prior is included. Defaults to TRUE.}
}
\value{
A float.
}
\description{
This computes the complete loss of the network on a timeseries.
If desired, the prior penalty can be included.
}
\examples{
net <- createNetwork(inputProbabilities=1, 5, 2, "homogeneous")
ts.multi <- simulateNetwork(net, 10, 50) # Creates 50 timeseries à 10 points
loss <- Loss(ts.multi, net)
}
\seealso{
\code{\link{LossGeneTotal}} \code{\link{LossGeneTimeseries}} \code{\link{LossPriorGene}}
}
