% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{LossGeneTimeseries}
\alias{LossGeneTimeseries}
\title{The loss on timeseries data for a single gene.}
\usage{
LossGeneTimeseries(ts.multi, net, g)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}
}
\value{
A float.
}
