% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{LossGeneTotal}
\alias{LossGeneTotal}
\title{The total loss of the network on a single gene.}
\usage{
LossGeneTotal(ts.multi, net, g)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}
}
\value{
A float.
}
