% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{LossPriorGene}
\alias{LossPriorGene}
\title{Penalty for single gene's regulatory set's complexity.}
\usage{
LossPriorGene(net, g)
}
\arguments{
\item{net}{A network, either threshold PBN or BoolNet net.}

\item{g}{The gene to be penalized.}
}
\value{
A float.
}
