% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{optimizeC}
\alias{optimizeC}
\title{Optimize probabilities for regulatory functions to be used.}
\usage{
optimizeC(ts.multi, net, g, prune = TRUE, thresh = 0.1)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{prune}{Whether or not to prune regulatory functions with low
probabilities assigned to them.}

\item{thresh}{The threshold below which to prune functions if prune
is TRUE.}
}
\value{
A network whose probabilities have been optimized.
}
\description{
Find best probabilities. Prune those with small values.
}
\details{
Using \code{\link{bestC}}, this finds the best probabilities. Afterwards, it
prunes all functions whose probabilities is below a user-defined threshold.
}
\seealso{
\code{\link{bestC}}
}
