% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{predict}
\alias{predict}
\title{Generate gene's next state given current state.}
\usage{
predict(net, g, state)
}
\arguments{
\item{net}{The network to be simulated.}

\item{g}{The gene whose value is to be predicted.}

\item{state}{The current state of the netwrok.}
}
\value{
The simulated value of the gene.
}
\description{
Generate gene's next state given current state.
}
