% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{reproduce}
\alias{reproduce}
\title{Function to reproduce all data and plots.}
\usage{
reproduce()
}
\description{
Running this script reproduces all the data and plots which have also been used
in the PDF report. The only thing which might differ should be the times taken
to do the inference, depending on your setup.
}
\examples{
\dontrun{reproduce()}
}
