% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setupPBN}
\alias{setupPBN}
\title{Generate a random TPBN and a timeseries simulated from it.}
\usage{
setupPBN(n = 20, k = 5, p = 0.01, inputProbabilities = c(0.5, 0.4, 0.1),
  topology = "homogeneous", gamma = 2.5, num.timepoints = 10,
  num.experiments = 20, seed = 111, verbal = FALSE)
}
\arguments{
\item{n}{Size of the network.}

\item{k}{The number of inputs per regulatory function for each gene, if homogeneous
topology is used.}

\item{p}{The probability of a perturbation.}

\item{inputProbabilities}{Probabilities for creating one, two, three, ... regulatory
functions for a gene.}

\item{topology}{The topology to be used. Can be "homogeneous" or "scale_free".}

\item{gamma}{The exponent for the power law if topology = "scale_free".}

\item{num.timepoints}{The number of time points per timeseries generated.}

\item{num.experiments}{The number of timeseries to generate.}

\item{seed}{The random seed to use.}
}
\value{
A list containing the generated network and a list of timeseries simulated
  from it.
}
\description{
Creates a network with the desired characteristics and simulates a list of timeseries
from it which can be used to test the network inference algorithm against.
}
\examples{
setupPBN(n=5, k=3, topology="scale_free", gamma=2.5)
}
\seealso{
\code{\link{transformNetwork}} \code{\link{simulateNetwork}}
}
