% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{transformNetwork}
\alias{transformNetwork}
\title{Turns BoolNet network into threshold PBN.}
\usage{
transformNetwork(net, p = 0.01)
}
\arguments{
\item{net}{A network generated by \code{\link{createNetwork}}.}

\item{p}{Noise strength.}
}
\value{
A threshold PBN.
}
\description{
Takes a BoolNet network and add threshold parameters as well
  associated logic functions to it.
}
\examples{
\dontrun{transformNetwork(net)}
}
