% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{A}
\alias{A}
\title{Add threshold parameters to network.}
\usage{
A(net)
}
\arguments{
\item{net}{The network to be transformed.}
}
\value{
The network with associated parameters.
}
\description{
Threshold parameters determine whether an input will have a positive or
negative relationship with the target genes.
}
