% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{aToFunction}
\alias{aToFunction}
\title{Turn threshold parameters into Boolean functions.}
\usage{
aToFunction(net)
}
\arguments{
\item{net}{A network which has threshold parameters.}
}
\value{
A Network with logic functions associated.
}
\description{
Take a network with a set of threshold parameters and compute the correct
Boolean functions belonging to these parameters.
}
\examples{
net <- generateRandomNKNetwork(5, 2, "homogeneous")
net <- A(net)
net <- aToFunction(net)
}
