% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{bestA}
\alias{bestA}
\title{Find best threshold parameters for regulatory function.}
\usage{
bestA(ts.multi, net, g, i)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{i}{The index of the regulatory function.}
}
\value{
A vector containing the best threshold parameters found.
}
\description{
des
}
\details{
See also \code{\link{bestPartialA}}
}
