% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{bestC}
\alias{bestC}
\title{Find best probabilities for the regulatory functions for single gene.}
\usage{
bestC(ts.multi, net, g)
}
\arguments{
\item{ts.multi}{List of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}
}
\value{
A vector containing the probabilities optimizing the loss function.
}
\description{
des
}
\details{
det
}
