% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize-PBN.R
\name{bestPartialA}
\alias{bestPartialA}
\title{Find the best threshold parameters for a single gene regulatory
  function, holding everything else constant.}
\usage{
bestPartialA(ts.multi, net, g, i, subset)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}

\item{i}{The index of the specific regulatory function for the optimization.}

\item{subset}{The subset of parameters over which to optimize.}
}
\value{
A vector containing the best threshold parameters found.
}
\description{
des
}
\details{
det
}
