% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{createNetwork}
\alias{createNetwork}
\title{Create random threshold PBN.}
\usage{
createNetwork(inputProbabilities = 1, ...)
}
\arguments{
\item{inputProbabilities}{List of probabilities determining how many
regulatory sets there are per gene.}

\item{...}{Other parameters, used by BoolNet.}
}
\value{
A network.
}
\description{
This first uses its parameters to let BoolNet create a network with the
desired nodes, number of edges, topology, etc.
It then creates adds all the relevant parameters to make it usable as a
TPBN.
}
\examples{
net <- createNetwork(inputProbabilities=1, 5, 2, topology="homogeneous")
}
\seealso{
\code{\link{transformNetwork}} \code{\link{setupPBN}}
}
