% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{Loss}
\alias{Loss}
\title{Loss function for time series.}
\usage{
Loss(ts.multi, net, prior = TRUE)
}
\arguments{
\item{ts.multi}{List of timeseries.}

\item{net}{A PBN.}

\item{prior}{whether or not the prior should be included. Defaults to TRUE.}
}
\value{
A float.
}
\description{
Returns the loss of the network on the timeseries.
}
\details{
Stuff.
}
