% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{LossGeneTimeseries}
\alias{LossGeneTimeseries}
\title{Computes the loss for a gene over a list of timeseries.}
\usage{
LossGeneTimeseries(ts.multi, net, g)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}
}
\value{
A float.
}
\description{
Makes use of \code{\link{LossGene}}, which computes the loss for only a
single timeseries.
}
\examples{
net <- createNetwork(inputProbabilities=1, 5, 2, "homogeneous")
ts.multi <- simulateNetwork(net, 10, 50) # Creates 50 timeseries à 10 points
loss <- LossGeneTimeseries(ts.multi, net, 1)
}
