% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{LossGeneTotal}
\alias{LossGeneTotal}
\title{Compute the total loss for a gene.}
\usage{
LossGeneTotal(ts.multi, net, g)
}
\arguments{
\item{ts.multi}{A list of timeseries.}

\item{net}{A network.}

\item{g}{A gene.}
}
\value{
A float.
}
\description{
This includes both the predictions on the timeseries as well as the penalty
for its complexity.
}
\examples{
net <- createNetwork(inputProbabilities=1, 5, 2, "homogeneous")
ts.multi <- simulateNetwork(net, 10, 50) # Creates 50 timeseries à 10 points
loss <- LossGeneTotal(ts.multi, net, 1)
}
