% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-network.R
\name{simulateNetwork}
\alias{simulateNetwork}
\title{Simulate a threshold PBN.}
\usage{
simulateNetwork(net, Tmax, Emax, start = "random", start.values,
  fix.genes = -1, fix.values = -1)
}
\arguments{
\item{net}{The network from which the timeseries is generated.}

\item{Tmax}{The number of time points per timeseries.}

\item{Emax}{The number of timeseries.}

\item{start}{If "random", gene values are initialized randomly.}

\item{start.values}{If start is not "random" these start values are used.}

\item{fix.genes}{Vector of genes to be kept fixed at a certain level.}

\item{fix.values}{Vector of values at which genes should be fixed.}
}
\value{
A list of timeseries.
}
\description{
From a given starting configurations, simulate the network Emax times for
Tmax steps in every single timeseries.
}
\examples{
net <- createNetwork(inputProbabilities=1, 5, 2, "homogeneous")
ts.multi <- simulateNetwork(net, 10, 50) # Creates 50 timeseries à 10 points
}
